+++ src/test/Shell.h
@@ -1,11 +1,7 @@
-#include <fstream>
+#pragma once
+
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <sstream>
 #include <string>
-#include <string_view>
-#include <vector>
 
 std::string exec(const std::string cmd) {
     auto pPipe = ::popen(cmd.c_str(), "r");
